pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["Hide <hshidara@ucdavis.edu>"]

[tool.poetry.dependencies]
python = "^3.10"
argon2-cffi = "^21.3.0"
Django = "3.2.9"
requests = "^2.27.1"
celery = "^5.2.6"
django-environ = "^0.8.1"
dj-database-url = "^0.5.0"
whitenoise = "^6.0.0"
django-hosts = "^5.1"
django-crispy-forms = "^1.14.0"
crispy-bootstrap5 = "^0.6"
django-celery-beat = "^2.2.1"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.11.0"
django-debug-toolbar = "^3.2.4"
django-extensions = "^3.1.5"
psycopg2 = "^2.9.3"
django-allauth = "^0.50.0"
django-anymail = "^8.5"
django-redis = "^5.2.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
